Set ulimit before our trap
We set a ERR trap in the compile.sh script to catch all failures, which
means when the ulimit fails, the entire build fails.

There seem to be random build boxes that don't allow you to set ulimit.
It is OK if setting it fails as we only use it to introspect core dumps
(hopefully rarely).

Setting the ulimit before the trap should fix this