PHPC-144: MO first-user changes
MO now creates an initial user that collides with the user and
certificate we used for this.
MO really needs to generate a different internal user as most people
will do as we did - take the user+cert from mongo jstests - rather then
mucking around and create the chain ourself.
To work around this, we create a bogus user (to trigger MO to create
its user) but use the internal MO user as it is the one we want