PHPC-450: failReceivedGetmore does not support getMore commands
The fail point used in this test does not yet support 3.2's getMore command, so this must be run on 3.0 (or earlier). See: SERVER-21831

This change also splits the test into query/command variants, corrects the title (which was apparently copied from cursor-getmore-004.phpt), and removes the use of data fixtures (small documents inserted within the test work just fine for its purposes).