Workaround for issue #133
When doing async connections we can't properly determine if we
should connect over IPv6 or IPv4. PHP will unfortunately default
to connecting over IPv6 (when compiled in support, with minor
runtime check), while mongod does not support IPv6 by default.

When then creating the IPv6 connection asyncronously, PHP hits
EINPROGRESS and will happily return the socket as valid.

There currently is no way to configure these functions to prefer IPv4.
The best we can do for now is to not connect asyncronously as then
we'll properly discover that the IPv6 socket failed, and then continue
through the list of addrinfo and next pickup the IPv4 address which
works.