PHPC-553: Check for connection exceptions in exec and SS methods
executeCommand(), executeQuery(), and selectServer() did not check for a global exception after executing via libmongoc, which meant that connection and SSL exceptions might end up as the previous exception where they could be easily ignored.

This makes these methods consistent with executeBulkWrite(), which happens to be used in our SSL tests where we check for such exceptions.