PHPC-544: Consult SIZEOF_ZEND_LONG for 64-bit integer support
Define phongo_long and SIZEOF_PHONGO_LONG in compatibility header. On PHP 7, this will ensure we use zend_long and SIZEOF_ZEND_LONG instead of relying on long and SIZEOF_LONG.

Additionally, this adds an else condition to report a build error if the architecture is neither 32-bit nor 64-bit.