Use both mongodb and openssl extension paths for CHECK_LIB
The ldap extensions' config.w32 uses PHP_LDAP as its search path, so we'll continue to start with PHP_MONGODB as a rule of thumb. Since we currently declare a dependency on the OpenSSL extension, it makes sense to also include it in our search path in case it was customized at build time.

Based on the openssl extension's config.w32, crypt32.lib is a system library and needs no extra search path.