PHPC-1092: Don't reset hash on successive get_properties calls
This removes a call to zend_hash_clean() on successive calls to a BSON object's get_properties handler, which can lead to an infinite loop when iterating the object's properties with foreach on PHP 7.x.

For historical reference, the original call to zend_hash_clean() was introduced in ad1f3f049f5c101a8b4f89f13e6384bbefb6ef6c for #607 (PHPC-939).