PHPC-1180: Do not require weak cert validation for STANDALONE_SSL
Per my original comment in the ticket, this URI option could conflict with older SSL-specific tests that relied on peer verification using PEM files. We'll still need to investigate those tests later, but it turns out this option wasn't needed for other tests to connect to an SSL server as previously assumed.