Use aggregate with $out in executeReadWriteCommand tests
MongoDB 4.0.x (and possibly earlier) happens to require an $out stage for aggregate if a writeConcern option would be used. While this test doesn't use a write concern, it could inherit one from MONGODB_URI.

As for findAndModify, that is only considered a write command (does not accept a readConcern option).