PHPC-2146: Refactor typemap struct and BSON encoding/decoding of zvals (#1369)
* Avoid re-parsing bson when converting to zval

A lot of occurrences start out with a bson_t*, then call bson_get_data on it to pass it to php_phongo_bson_to_zval_ex. This however creates a new bson reader and creates a new bson_t* from the given data. This is unnecessary and we should pass the original bson_t* in these instances.

* Refactor typemap struct

* Extract object initialisation from document visitor

* Refactor document and array visitors

* Fix wrong signature in array visitor declaration

* Rename ce member for consistency

* Add exception thrown comment

* Fix clang-format