Ensure ZEND_POW yields correct results for Int64
The previous usage of pow() was problematic, as it returns a double which results in a loss of precision compared to int64_t. The new logic implements exponentiation by squaring to ensure that we're returning int64_t values where possible while guarding against overflow and reverting to PHP's default behaviour, which is to return a double. Since PHP's own implementation of this algorithm only supports int32_t on 32-bit platforms, we have to reimplement it to support the full int64_t range on 32-bit platforms.