PHPC-2214: Int64 improvements (#1417)
* PHPC-1949: Allow construction of Int64 objects

* PHPC-2213: Support casting of Int64 objects

* PHPC-2212: Support overloaded operators for int64

* Conditionally declare zend_result type on PHP < 8

* Fix exception copypasta

* Guard against division by zero for modulo operation

* Refactor do_operator handler for Int64

* Support bitwise operators for Int64

* Add tests for inc/dec operations on Int64

* Ensure ZEND_POW yields correct results for Int64

The previous usage of pow() was problematic, as it returns a double which results in a loss of precision compared to int64_t. The new logic implements exponentiation by squaring to ensure that we're returning int64_t values where possible while guarding against overflow and reverting to PHP's default behaviour, which is to return a double. Since PHP's own implementation of this algorithm only supports int32_t on 32-bit platforms, we have to reimplement it to support the full int64_t range on 32-bit platforms.

* Incorporate review feedback

* Handle overflow on arithmetic operations

* Expand tests for bitwise operations

* Add test for invalid constructor values for Int64

* Add operator tests for ZEND_BOOL_NOT and ZEND_BOOL_XOR

* Add clarifying comments for Int64 arithmetic operations

* Support casting Int64 objects to float

* Use PHP_INT_MIN constant instead of negation

* Harden regular expression to match scientific notation

* Create int64 instances directly in tests

* Fix return types of increment/decrement operators

* Add test for casting to float