PHPC-2638 and PHPC-2630: Bump to libmongoc 2.1.2 and libmongocrypt 1.16.0 (#1868)
* PHPC-2638: Bump libmongoc to 2.1.2

* PHPC-2630: Bump libmongocrypt to 1.16.0

This pulls in changes from MONGOCRYPT-845, which are necessary to produce a static build with both libmongoc 2.1.x and libmongocrypt (due to a conflict in the "mlib" headers).

* Update SBOM for libmongoc 2.1.2 and libmongocrypt 1.16.0

* Update error expectations for libmongoc 2.1

* Update libmongocrypt mlib path for PECL packaging

* Bump libmongoc and libmongocrypt versions for system library testing

* Cross-compile 32-bit Windows builds using 64-bit toolset

This should avoid hitting memory limits when using the 32-bit toolset, which is relevant for libmongoc sources that make heavy use of the BSON DSL API (so many macros).